fix: Remove EtcdConfigController
