+++ src/ai/backend/manager/config/unified.py
@@ -1,8 +1,9 @@
-from typing import Awaitable, Callable
+import asyncio
+from typing import Awaitable, Callable, Optional
 
-from ai.backend.manager.config.watchers.types import AbstractConfigController
+from ai.backend.manager.config.watchers.etcd import EtcdConfigWatcher
 
-from .loader.etcd_loader import LegacyEtcdLoader
+from .loader.legacy_etcd_loader import LegacyEtcdLoader
 from .loader.types import AbstractConfigLoader
 from .local import ManagerLocalConfig
 from .shared import ManagerSharedConfig
@@ -14,20 +15,34 @@ class ManagerUnifiedConfig:
     local: ManagerLocalConfig
     shared: ManagerSharedConfig
     local_config_loader: AbstractConfigLoader
-    etcd_config_loader: LegacyEtcdLoader
+    legacy_etcd_config_loader: LegacyEtcdLoader
+    etcd_watcher: EtcdConfigWatcher
 
-    config_controllers: list[AbstractConfigController]
+    _etcd_watcher_task: Optional[asyncio.Task[None]]
 
     def __init__(
         self,
         local: ManagerLocalConfig,
         shared: ManagerSharedConfig,
         local_config_loader: AbstractConfigLoader,
-        shared_config_loader: LegacyEtcdLoader,
-        config_controllers: list[AbstractConfigController] = [],
+        etcd_config_loader: LegacyEtcdLoader,
+        etcd_watcher: EtcdConfigWatcher,
     ) -> None:
         self.local = local
         self.shared = shared
         self.local_config_loader = local_config_loader
-        self.etcd_config_loader = shared_config_loader
-        self.config_controllers = config_controllers
+        self.legacy_etcd_config_loader = etcd_config_loader
+        self.etcd_watcher = etcd_watcher
+
+    async def _run_watcher(self) -> None:
+        async for event in self.etcd_watcher.watch():
+            # TODO: Handle all etcd_loader.load() here
+            pass
+
+    def start(self) -> None:
+        self._etcd_watcher_task = asyncio.create_task(self._run_watcher())
+
+    def stop(self) -> None:
+        if self._etcd_watcher_task:
+            self._etcd_watcher_task.cancel()
+            self._etcd_watcher_task = None
