+++ src/ai/backend/manager/config/watchers/types.py
@@ -7,13 +7,3 @@ class AbstractConfigWatcher(ABC):
     @abstractmethod
     async def watch(self) -> Mapping[str, Any]:
         raise NotImplementedError
-
-
-class AbstractConfigController(ABC):
-    @abstractmethod
-    async def start(self) -> None:
-        raise NotImplementedError
-
-    @abstractmethod
-    async def stop(self) -> None:
-        raise NotImplementedError
